Remove tools/build.py (#2865)
Testing regression: ASAN build removed.
