--- build_extra/rust/empty_crate.rs
@@ -1,2 +1,0 @@
-// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-fn main() {}
