--- build_extra/rust/get_rustc_info.cmd
@@ -1,2 +1,0 @@
-REM Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-@"%PYTHON_EXE%" "%~dpn0.py" %*
