--- deno_typescript/BUILD.gn
@@ -1,22 +1,0 @@
-import("//build_extra/rust/rust.gni")
-
-rust_rlib("deno_typescript") {
-  source_root = "lib.rs"
-  generated_source_dir = "."
-  extern = [
-    {
-      label = "../core:deno"
-      crate_name = "deno"
-      crate_type = "rlib"
-    },
-    {
-      label = "$rust_build:serde_derive"
-      crate_name = "serde_derive"
-      crate_type = "proc_macro"
-    },
-  ]
-  extern_rlib = [
-    "serde_json",
-    "serde",
-  ]
-}
